Removed React and TypeScript
Reason why I removed React and TypeScript. I am currently the only developer, the purpose of this project was for me to study these Frameworks. However, I've made huge life changes, and I need this project to be done by the end of this year with one Minecraft Development Tool to demonstrate the launcher's...  launch and some use of API.

Compiler fails due to the way the props are passed in the attributes via React's render() and I'm incapable in resolving this issue alone as I'm not qualified for React (yet). This Framework is beneficial for performance, but it delays the overall development and is unnecessary for the project's size.

TypeScript is removed, not on purpose, I liked it but couldn't be bothered to install it as I rebuilt the app from scratch. I see no point in installing it as previously mentioned, mainly due to the project's relatively small size (and I don't see it being scaled in the future)

I would like to use this project to strengthen my JavaScript skills, and I don't want to distract myself with alien to me code, hence why I only rely on Electron alone for future development of the Minecraft-DT2.

Improvements:
Many functions where simplified. (especially those responsible for validating and importing development tools)

Development Tools folder is now located in root directory.

Main functions (responsible for importing development tools) where moved to the preload.js. Any other function related specifically for manipulating/initializing the renderer are located in index.js and in other external scripting files, all found in "src" directory.

Last major change: <help-tag> functionality has changed. You are no longer able to initialize the help-tag contents based on node's attribute data. Here is a new neat method for defining node for help-tag:

<button onclick="alert('hello')">
  Click me
  <help-content name="display hello message">
    <p>My message <span class="warning">on click!</span>
  </help-content>
</button>

Parent <node> can only have one <help-content> child node.

Also, there are now two classes for <help-content>
<span class="warning"> and
<span class="danger">
Both used to change the colour of the message.

That's all for now.