AAAAAA had no wifi so don't bother looking for explanation as too much has happened
I'll try to make it short:

Basically, I have removed React from the launcher Minecraft-DT2. I do not believe that everything is finished, there might be mistakes and new features added, not that I remember. So lets just accept what we have to deal with as is. This application is not scaleful enough to require such description as it takes a week to fully understand it's data flow.

Another thing: I am currently working on first development tool - Command Creation Generator. There are two editors within this app. My goal is to finish this application first before I start polishing the launcher with planned download-tool API.

Current phase of Command Creation Generator - Design and Implementation. I am currently satisfied with the Sign's editor although, some features may be added if necessary. The main focus is on the command's text editor, I would like to implement a text highlighting system but it is way beyond my abilities therefore, I need to use a third party library.

Once desired design is implemented I will focus on functionality. The goal is to generate the working Minecraft creation, with two command blocks. This development-tool's goal should also be able to save and read it's projects. Hope that makes all clear. :)