Added new functions and fixed a bug at command creation generator development tool
+Added loadPrismLive() to restart prismLive
+Added containerGetCommands() at container.js (self explanatory)

Returns array of commands from the editor-container

+Added containerSetCommands(commands_array=[]) at container.js (self explanatory)

It restarts .prism-live with commands provided by an array.

+wysiwygCommandsObserver, It changes multiplier (a style property used for #commands > textarea) based on .zoomContainer's percentage (input value)

There was a bug, when #commands is scaled with zoom (style property), the max-width of the textarea has increased, allowing user to scale the textarea horizontally beyond the client's width.

There is another bug but it is impossible to reproduce with the desktop application. The desktop client cannot be scaled the way it is defined. However, launching the development tool's index.html file in a browser, user can scale the client content, this breaks the #commands textarea's minimal width. Hence why, either we can fix this or just don't bother by keeping the desktop app's initial scale to 100%.

+At last, I've updated the parameter's validation at "createInterval" library. I suggest moving this library into separate JavaScript file as it is also used in the main launcher's index.js (found at "./mdt/src/scripts" folder)