Still working on sorting algorithm
I have moved unrelated to the project's (Development Tool - Command Creation Generator) contribution files (./CCG-Progress) outside of main source directory

-/+ I moved object's 'Intervals' functionality to separate JavaScript file to organise some space. I found it to be very useful so I may use it often in other projects.

Development-tool: [Command Creation Generator]:

I mainly focused on creating the sorting algorithm with animations using the function containerGetCompiledCode() at ./container.js

I currently had no idea it will grow so large. Hence in the future, something must be done with the nesting functions to split it out. Organise it but for now, my main goal is to make it work.

There are stages to help understand the animation process

Stage1: Prepares code display for the algorithm by replacing strings that may become problematic

Stage2: Imports stripped from unwanted spacing code to speed up the algorithm (It also prevents any execution issues by replacing possible tabs with spaces)

Stage3: (The most confusion one) It's purpose is to identify, validate, and compress closed brackets. Closed brackets may exist only in strings as this sorting algorithm has no access to code's string tokens since they've been stripped.

There is a good reason I did this. Strings can be written for a message in minecraft, hence, we don't want to interfere with user's desired outcome.

The only reason I could use the sorting in property command. E.g. "command": "/tellraw bob"

But that is a hassle. I could do it but sadly too much work is required and I have other things that take priority.

As you can see the script is not finished as new command does not return every data from past conversions.

IMPORTANT: Also in stage 3, there are some comments for testing purposes, It misses one that I have recently discovered. The sorting algorithm gets stuck in a infinite loop when we provide "{},}" or "[],]" in code. It happens only when there is a punctuation "," after closed bracket, and is followed later by a closed bracket "]" or "}". Hence, THIS -> ",}" is an issue that is causing serious stress on the user's device. This must be fixed at all cost.