Fixing bugs
The bracket bug is fixed "[...] ,]" I forgot to highlight and remove the non-bracket punctuation, so not a biggie.

The function is now able to return compressed commands in a list with original quotes which where replaced temporary so that the algorithm does not interfere with them.

There was another, well... mistake, where updated code is returned... split with spaces using commas. E.g.
IN -> "/hello {
    bob,
    james,
    jerry
}"

OUT -> "/hello {,    bob,    ,james,    jerry, }"

It's fixed now.

Last mistake, I can't use errorLine to get the X position that will scroll the textarea horizontally to .error token. A stupid mistake, it supposed to be .error token, which is denoted as 'opened_bracket'. Fixed, works.

All now is to do a heavy test. I tried complex testing, but I feel that more is required in case I've missed something. I also worry on performance due to the frequent use of containerSetCode(). (PrismJS Live does not support restart function, not that I am aware of)