Popup GUI
We can now gather or inform user using new popups GUI.

Generate Creation button has it's own confirm popup.

We are now greeted with a loading screen before we see the app's content, it purpose is to wait until every necessary script is loaded and initialized at main.js

Added validation for parameter 'sign_instance' at getSingData() function.

Updated template script at container editor. I've decided to keep it as an example, so that the user can hopefully be able to see how can Minecraft's code be dynamically typed.

I removed getCaretIndex() as it's not being used and moveCaretTo() as well, I think (I forgot, there might be possibility of it being moved to more adequate script, which is the wysiwyg.js)