added 4 new functions
Not much today:

The current development allows us to getSignDataTag() which is the main focus. I managed to end up with 3 functions as I did not wanted to nest them, on the other hand, but in return this allows flexible access for specific needs, yey.

The improvement tip, since we NOW know the format of minecraft's dataTag object and ended up with getSignDataTagObject(), we could technically construct the same object instead of relying on getSignData()'s key 'display'.But since the software works and it does not slow down the software's performance, I'll leave it open for future development.

The next goal is to also include minecraft's key 'run_command' in the 'dataTag'. After that, we can include valid commands from container-editor to the container's summoning command(s) (does not exist yet, and it's far ahead but close enough to start thinking about the commands structure)