Included clickEvent key to DataTag
'clickEvent' key is required for sign's command(s) execution on right click event. So, I have created an object that defines 'clickEvent' (Hence clickEvent's object's additional keys: 'action' and 'value')

Do note that final value for 'value' key (which is a command) is a bit manipulated. There is no need for character '/' at the start of the command. Also, unnecessary spaces are stripped out before and after the command string.