Added two functions getImportCreationCommand() and getSummonContainerCommand()
getImportCreationCommand() and getSummonContainerCommand() are are stated.

We can now convert user's input and convert it to executable summoning commands for creation's inclusion in Minecraft.

However, the project is not finished yet! Because we still have to replace variables !(creation_programmer), !(creation_version) and !(creation_name) in both summoning commands.

After that, we may want to prompt user to save the generated code in a .html file so that their creation can be easily shared with the world.

Another essential feature is to inform user of both summoning command's limit, this can be easily added in the generated .html file.

Also note that user is able to generate importCreationCommand() with more than there are allocated commandBlocks (112 commands is the limit) I forgot to include validation, I'll do it next time and then update the version with documented progress.